updated Waiter Interface

+++ src/main/java/com/project/restaurantOrderingManagement/service/tableService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class tableService {
@@ -20,12 +21,18 @@ public class tableService {
         return tableRepo.findAllByStatus(status);
     }
 
-    public table addTable(String code, table table) {
+    public table addTable(table table) {
         return tableRepo.save(table);
     }
 
     public void deleteTable(int tableNo) {
-        tableRepo.deleteById(tableNo);
+        Optional<table> existingItem = tableRepo.findById(tableNo);
+        if(existingItem.isPresent()) {
+            tableRepo.deleteById(tableNo);
+        }
+        else{
+            throw new RuntimeException("Table not found with ID: " + tableNo);
+        }
     }
     //update waiter alone
     public table updateTableByWaiter(int tableNo, table table) {
